Settings page changes, switch on Map screen
Settings page now sports a drop-down menu because slider has depreciated. I'll be checking consistently to see if it's fixed in another update in the future. For now, the dropdown operates the same. It's rough, robust, and not pretty (ha)

furthermore, there's a switch added to the map screen to swap between plane and satellite mode.